rename the exponent details folder to impl
