+++ include/ccmath/detail/exponential/impl/log2_data.hpp
