+++ include/ccmath/detail/exponential/impl/log2_float_impl.hpp
