+++ include/ccmath/detail/exponential/impl/log_data.hpp
