+++ include/ccmath/detail/exponential/log2.hpp
@@ -10,8 +10,8 @@
 
 #include "ccmath/detail/compare/isnan.hpp"
 #include "ccmath/detail/compare/signbit.hpp"
-#include "ccmath/detail/exponential/details/log2_double_impl.hpp"
-#include "ccmath/detail/exponential/details/log2_float_impl.hpp"
+#include "ccmath/detail/exponential/impl/log2_double_impl.hpp"
+#include "ccmath/detail/exponential/impl/log2_float_impl.hpp"
 
 #include <limits>
 #include <type_traits>
@@ -46,6 +46,7 @@ namespace ccm
 			else { return std::numeric_limits<T>::quiet_NaN(); }
 		}
 
+		// We can not handle long double at this time due to problems with long double being platform dependent with its bit size.
 		if constexpr (std::is_same_v<T, float>) { return ccm::internal::log2_float(num); }
 		else { return ccm::internal::log2_double(num); }
 	}
